Seperate scene construction and scene start
This way, Scenes can be constructed and queued to be switched on the next frame, rather than in the middle of one. Now, construction should ideally only be used for resource allocation and initialization. `start()` is when you can start messing with other things.

Also, experiment with different physics timestep stuff